+++ packages/@webex/plugin-meetings/src/locus-info/parser.ts
@@ -397,6 +397,7 @@ export default class Parser {
     const {isLoci} = Parser;
     // @ts-ignore
     const setStatus = (status) => {
+      // @ts-ignore
       this.status = status;
     };