+++ packages/@webex/plugin-meetings/src/media/properties.ts
@@ -225,7 +225,7 @@ export default class MediaProperties {
       return Promise.resolve();
     }
 
-    return new Promise((resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       let timer;
 
       const iceListener = () => {