+++ packages/@webex/plugin-meetings/src/reactions/reactions.type.ts
@@ -5,7 +5,8 @@ export type EmoticonData = {
 };
 
 export type SkinTone = EmoticonData;
-export type Reaction = ReactionData & {
+// @ts-ignore
+export type Reaction = EmoticonData & {
   tone?: SkinTone;
 };